Remove recover_backlog for recover_by_checkpoint instead
